Move ROOT_USER to a separate file, fixes #2807
