+++ __tests__/constants.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import {getPathKey, isRootUser} from '../src/constants.js';
+import {getPathKey} from '../src/constants.js';
 
 test('getPathKey', () => {
   expect(getPathKey('win32', {PATH: 'foobar'})).toBe('PATH');
@@ -9,10 +9,4 @@ test('getPathKey', () => {
   expect(getPathKey('win32', {})).toBe('Path');
   expect(getPathKey('linux', {})).toBe('PATH');
   expect(getPathKey('darwin', {})).toBe('PATH');
-});
-
-test('isRootUser', () => {
-  expect(isRootUser(null)).toBe(false);
-  expect(isRootUser(1001)).toBe(false);
-  expect(isRootUser(0)).toBe(true);
 });
