+++ src/constants.js
@@ -90,16 +90,3 @@ export function getPathKey(platform: string, env: Env): string {
 
   return pathKey;
 }
-
-function getUid(): ?number {
-  if (process.platform !== 'win32' && process.getuid) {
-    return process.getuid();
-  }
-  return null;
-}
-
-export const ROOT_USER = isRootUser(getUid());
-
-export function isRootUser(uid: ?number): boolean {
-  return uid === 0;
-}
