+++ src/util/user-home-dir.js
@@ -1,6 +1,8 @@
 /* @flow */
+
+import ROOT_USER from './root-user.js';
+
 const path = require('path');
-const {ROOT_USER} = require('../constants');
 
 const userHomeDir = (process.platform === 'linux' && ROOT_USER) ?
   path.resolve('/usr/local/share') :
